feat(CombineGWASWorkflowAttritionTableLogic): Reorganized files into AttritionTableWrapper directory
