+++ src/Analysis/GWASV2/Shared/AttritionTableWrapper/AttritionTable.jsx
