+++ src/Analysis/GWASV2/Shared/AttritionTableWrapper/AttritionTableWrapper.jsx
