+++ src/Analysis/GWASV2/Shared/AttritionTableWrapper/ChartIcons/BarChart.jsx
